Fixed a warning do to the treadID being a uint32_t

+++ Saturn/src/Saturn/Debug/Instrumentor.h
@@ -13,7 +13,7 @@ namespace Saturn
 	{
 		std::string Name;
 		long long Start, End;
-		uint32_t ThreadID;
+		size_t ThreadID;
 	};
 
 	struct InstrumentationSession
